LayoutTests/http/tests/security/contentSecurityPolicy/1.1/referrer-always-http-https.html

<!DOCTYPE html>
<html>
<head>
    <script src="/js-test-resources/js-test.js"></script>
    <script src="../resources/referrer-test.js"></script>
</head>
<body>
    <script>
        description("Test that 'unsafe-url' works correctly from HTTP to HTTPS.");
        expectFullReferrer(ReferrerTest.UNSAFE_URL, ReferrerTest.HTTP, ReferrerTest.HTTPS);
    </script>
</body>
</html>
